Include <simdjson> directly, move document_parser_callbacks to top level
