+++ src/arm64/bitmanipulation.h
@@ -1,8 +1,7 @@
 #ifndef SIMDJSON_ARM64_BITMANIPULATION_H
 #define SIMDJSON_ARM64_BITMANIPULATION_H
 
-#include "simdjson/common_defs.h"
-#include "simdjson/portability.h"
+#include <simdjson.h>
 
 #ifdef IS_ARM64
 
@@ -15,9 +14,10 @@ namespace simdjson::arm64 {
 // but the algorithms do not end up using the returned value.
 // Sadly, sanitizers are not smart enough to figure it out. 
 __attribute__((no_sanitize("undefined"))) // this is deliberate
-#endif
+#endif // _MSC_VER
 /* result might be undefined when input_num is zero */
 really_inline int trailing_zeroes(uint64_t input_num) {
+
 #ifdef _MSC_VER
   unsigned long ret;
   // Search the mask data from least significant bit (LSB) 
@@ -26,8 +26,9 @@ really_inline int trailing_zeroes(uint64_t input_num) {
   return (int)ret;
 #else
   return __builtin_ctzll(input_num);
-#endif// _MSC_VER
-}
+#endif // _MSC_VER
+
+} // namespace simdjson::arm64
 
 /* result might be undefined when input_num is zero */
 really_inline uint64_t clear_lowest_bit(uint64_t input_num) {
@@ -54,8 +55,7 @@ really_inline int hamming(uint64_t input_num) {
    return vaddv_u8(vcnt_u8((uint8x8_t)input_num));
 }
 
-really_inline bool add_overflow(uint64_t value1, uint64_t value2,
-                                uint64_t *result) {
+really_inline bool add_overflow(uint64_t value1, uint64_t value2, uint64_t *result) {
 #ifdef _MSC_VER
   // todo: this might fail under visual studio for ARM
   return _addcarry_u64(0, value1, value2,
@@ -70,20 +70,19 @@ really_inline bool add_overflow(uint64_t value1, uint64_t value2,
 #pragma intrinsic(_umul128) // todo: this might fail under visual studio for ARM
 #endif
 
-really_inline bool mul_overflow(uint64_t value1, uint64_t value2,
-                                uint64_t *result) {
+really_inline bool mul_overflow(uint64_t value1, uint64_t value2, uint64_t *result) {
 #ifdef _MSC_VER
   // todo: this might fail under visual studio for ARM
   uint64_t high;
   *result = _umul128(value1, value2, &high);
   return high;
 #else
-  return __builtin_umulll_overflow(value1, value2,
-                                   (unsigned long long *)result);
+  return __builtin_umulll_overflow(value1, value2, (unsigned long long *)result);
 #endif
 }
 
-}// namespace simdjson::arm64
+} // namespace simdjson::arm64
 
-#endif //IS_ARM64
-#endif //  SIMDJSON_ARM64_BITMANIPULATION_H
+#endif // IS_ARM64
+
+#endif // SIMDJSON_ARM64_BITMANIPULATION_H
