+++ src/arm64/bitmask.h
@@ -1,11 +1,10 @@
 #ifndef SIMDJSON_ARM64_BITMASK_H
 #define SIMDJSON_ARM64_BITMASK_H
 
-#include "simdjson/portability.h"
+#include <simdjson.h>
 
 #ifdef IS_ARM64
 
-#include "simdjson/common_defs.h"
 #include "arm64/intrinsics.h"
 
 namespace simdjson::arm64 {
