+++ src/error.cpp
@@ -1,4 +1,4 @@
-#include "simdjson/error.h"
+#include <simdjson.h>
 #include <map>
 
 namespace simdjson {
