+++ src/haswell/implementation.h
@@ -1,11 +1,10 @@
-#ifndef __SIMDJSON_HASWELL_IMPLEMENTATION_H
-#define __SIMDJSON_HASWELL_IMPLEMENTATION_H
+#ifndef SIMDJSON_HASWELL_IMPLEMENTATION_H
+#define SIMDJSON_HASWELL_IMPLEMENTATION_H
 
-#include "simdjson/portability.h"
+#include <simdjson.h>
 
 #ifdef IS_X86_64
 
-#include "simdjson/implementation.h"
 #include "isadetection.h"
 
 namespace simdjson::haswell {
@@ -27,4 +26,4 @@ public:
 
 #endif // IS_X86_64
 
-#endif // __SIMDJSON_HASWELL_IMPLEMENTATION_H
+#endif // SIMDJSON_HASWELL_IMPLEMENTATION_H
