+++ src/haswell/stage1_find_marks.h
@@ -1,7 +1,7 @@
 #ifndef SIMDJSON_HASWELL_STAGE1_FIND_MARKS_H
 #define SIMDJSON_HASWELL_STAGE1_FIND_MARKS_H
 
-#include "simdjson/portability.h"
+#include <simdjson.h>
 
 #ifdef IS_X86_64
 
@@ -56,4 +56,5 @@ WARN_UNUSED error_code implementation::stage1(const uint8_t *buf, size_t len, do
 UNTARGET_REGION
 
 #endif // IS_X86_64
+
 #endif // SIMDJSON_HASWELL_STAGE1_FIND_MARKS_H
