+++ src/jsonioutil.cpp
@@ -1,4 +1,4 @@
-#include "simdjson/jsonioutil.h"
+#include <simdjson.h>
 #include <cstdlib>
 #include <cstring>
 #include <climits>
