+++ src/stage2_build_tape.cpp
@@ -1,11 +1,8 @@
+#include <simdjson.h>
 #include <cassert>
 #include <cstring>
-
-#include "simdjson/portability.h"
-#include "simdjson/common_defs.h"
 #include "jsoncharutils.h"
-#include "simdjson/inline/document.h"
-#include "inline/document_parser_callbacks.h"
+#include "document_parser_callbacks.h"
 
 using namespace simdjson;
