+++ src/westmere/intrinsics.h
@@ -6,6 +6,7 @@
 #include <intrin.h> // visual studio
 #else
 #include <x86intrin.h> // elsewhere
-#endif //  _MSC_VER
-#endif //  IS_X86_64
-#endif //  SIMDJSON_WESTMERE_INTRINSICS_H
+#endif // _MSC_VER
+#endif // IS_X86_64
+
+#endif // SIMDJSON_WESTMERE_INTRINSICS_H
